+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # 2.x branch
 ## 2.6 branch
+### 2.6.2
+* fixed serious bug for fancybox;
+* datepicker uses the `AssetHelper`;
+* removed the `Cache` class.
+
 ### 2.6.1
 * fixed bug for form inputs;
 * removed some useless methods.